+++ app/Http/Controllers/Admin/SemesterController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\StoreSemesterRequest;
 use App\Http\Requests\Admin\UpdateSemesterRequest;
 use App\Models\Semester;
+use App\Services\Contracts\AcademicYearServiceInterface;
 use App\Services\Contracts\SemesterServiceInterface;
 use Illuminate\Http\Request;
 
@@ -13,16 +14,27 @@ class SemesterController extends Controller
 {
     public function __construct(
         protected SemesterServiceInterface $semesterService,
+        protected AcademicYearServiceInterface $academicYearService,
     ) {}
 
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Semester::class);
+
         return $this->semesterService->index($request);
     }
 
     public function create()
     {
-        return view('admin.semesters.create');
+        $this->authorize('create', Semester::class);
+
+        $academicYears = $this->academicYearService->query()
+            ->get();
+
+        return view(
+            'admin.semesters.create',
+            compact('academicYears')
+        );
     }
 
     public function store(StoreSemesterRequest $request)
@@ -34,9 +46,17 @@ class SemesterController extends Controller
 
     public function edit(Semester $semester)
     {
+        $this->authorize('update', $semester);
+
+        $academicYears = $this->academicYearService->query()
+            ->get();
+
         return view(
             'admin.semesters.edit',
-            compact('semester')
+            compact(
+                'semester',
+                'academicYears'
+            )
         );
     }
 
@@ -52,6 +72,8 @@ class SemesterController extends Controller
 
     public function destroy(Semester $semester)
     {
+        $this->authorize('delete', $semester);
+
         return $this->semesterService->delete(
             $semester
         );
